feat: add comprehensive improvements across frontend and backend
- Add FunASR realtime service for speaker diarization
- Add GraphRAG panel component and debug page
- Add test suite with pytest (test_api, test_asr_base, test_asr_factory, test_config, test_models)
- Add startup scripts (start_all.sh, start_funasr.sh)
- Add architecture diagram and documentation
- Improve LLM analyzer with enhanced chapter detection
- Fix audio streaming endpoint and upload response handling
- Add processing status tracking module
- Update .gitignore to exclude workspace output and .claude

+++ meeting-voice-assistant/backend/app/core/models/__init__.py
@@ -43,8 +43,8 @@ class ActionItem:
 class Chapter:
     """章节"""
     title: str
-    start_time: int
-    end_time: int
+    start_time: float
+    end_time: float
     summary: str
     speaker_summaries: List[SpeakerSummary] = field(default_factory=list)
     decisions: List[Decision] = field(default_factory=list)
